Fix typo

+++ trunk/onboard/fet.c
@@ -83,7 +83,7 @@ void fetCreateSine(void) {
 	a = M_PI * 2.0f * i / FET_SERVO_RESOLUTION;
 
 	// third order harmonic injection
-	fetSine[i] = (sinf(a) + sin(a*3.0f)/6.0f) * (2.0f/sqrtf(3.0f)) * (float)fetPeriod / 2.0f;
+	fetSine[i] = (sinf(a) + sinf(a*3.0f)/6.0f) * (2.0f/sqrtf(3.0f)) * (float)fetPeriod / 2.0f;
     }
 }
